[SPARK-13485] Cleanup dependencies between SQLContext and SparkSession
